requirements.txt
openapi-generator-cli
urllib3
python_dateutil
pydantic
typing-extensions
python-dotenv
setuptools
pytest
